# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EVO Platform API",
    "version": "3.0.0",
    "description": "EVO Platform by NuevaCore - Multi-Cloud Security & Cost Management Platform. Supports AWS and Azure."
  },
  "servers": [
    {
      "url": "https://api.evo.nuevacore.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/functions/security-scan": {
      "post": {
        "summary": "Run security scan (23 scanners, 170+ checks)",
        "operationId": "security_scan",
        "tags": [
          "Security"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "type": "object",
                      "description": "Response data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input parameters"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Not found - Resource does not exist"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/security_scan_request"
              }
            }
          }
        }
      }
    },
...
    "/api/functions/azure-security-scan": {
      "post": {
        "summary": "Azure security scan",
        "operationId": "azure_security_scan",
        "tags": [
          "Azure",
          "Security"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "type": "object",
                      "description": "Response data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input parameters"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Not found - Resource does not exist"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "security_scan_request": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "credentialId": {
            "type": "string",
            "format": "uuid"
          },
          "regions": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-z]{2}-[a-z]+-\\d$"
            }
          },
          "scanTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "frameworks": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CIS",
                "WELL_ARCHITECTED",
                "PCI_DSS",
                "NIST",
                "LGPD",
                "SOC2"
              ]
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "AWS Cognito JWT token from Authorization header"
      }
    }
  },
...
}